pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "product-review-sentiment-mlops"
version = "0.1.0"
description = "End-to-end e-commerce review sentiment analyzer with MLOps practices."
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.4.0",
  "pandas>=2.1.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.5.0",
  "datasets>=3.0.0",
  "matplotlib>=3.8.0",
  "joblib>=1.4.0",
  "mlflow>=2.15.0",
  "prometheus-client>=0.20.0",
  "python-json-logger>=2.0.7",
  "httpx>=0.27.0",
  "PyYAML>=6.0.0",
  "psutil>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "dvc>=3.50.0",
]
airflow = [
  "apache-airflow>=3.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ml*", "apps*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
